refactor(connect): move onCallFirmwareUpdate to separate file
